+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 **Added**
 
 - Dialect for JSONC
+- [Lint source](https://codemirror.net/docs/ref/#lint.LintSource) for JSONC